Add multiple parameters (without updating the context)
